Define output value for the cicd_pipeline stack, containing the ARN of the build project's role
This is necessary because the build project will need to have ECR access in order to push images during the build process.
Since this access will be given through a policy defined in another stack, and that policy need to reference the role used
by the build project, it is necessary to pass the ARN value between stacks. That will be done by storing the output in a JSON file.